backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "starlette>=0.27.0",
    "sqlmodel>=0.0.18",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.1.0",
    "bcrypt>=4.1.2",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
]
